Move testing code to main hellsemble
