--- testing/test_hellsemble/__init__.py
@@ -1,1 +1,0 @@
-from .hellsemble import Hellsemble  # noqa: F401
